Rewrite allluxio changing mountpoint test case
